+++ Vungle/CHANGELOG.md
@@ -1,4 +1,8 @@
 ## Changelog
+  * 6.9.0.0
+    * This version of the adapters has been certified with Vungle 6.9.0 and MoPub SDK 5.14.0.
+    * New callback `onAdViewed` is now used and mapped for MoPub's `onAdImpression`.
+
   * 6.8.1.0
     * This version of the adapters has been certified with Vungle 6.8.1 and MoPub SDK 5.14.0.
     * Added support for Android 11.